test: keep the disc layout tests in the one file they live in
